Handle errors and print failed item IDs at the end
If there is an error when getting the entity data, or while processing
the data, report the error but resume with the next item. At the end of
the whole run, report all the problems at once so users don’t have to
scroll through tons of output (mwbot errors can be rather verbose).