Move most of namescript.js to namescript-lib.js
The plan is that namescript-lib will contain the functionality wrapped
in an IIFE, exposed via the “namescript” global, and namescript.js will
use that. But I’m afraid that if I do all of those steps at once – move
to a different file *and* reindent for IIFE – the lines will just look
like new lines in e. g. a `git blame`. Therefore I’m first just moving
all of the code, temporarily leaving a broken program, and then I’ll fix
it up in a separate commit.