Rename namescript.js to namescript-cli.js
This is the CLI entry point; a separate one for the browser will also be
added, and then a new namescript.js will automatically detect which one
to load, for convenience and backwards compatibility.

Also load namescript-lib.js, so that this version should be functional
again.

Also add a catch to the final main() call to avoid an “unhandled promise
rejection” warning.