Display API errors in the browser
The async chain is broken at add() in the browser, so at that point we
need to do something with the API error if it happens. mw.Api() resolves
the jQuery Promise with four arguments, but native Promise only accepts
one argument and throws the others away, so to get the full result
instead of just the error code the jQuery→native conversion needs to get
a bit more uglier.

I’m not sure what this will do with errors that don’t have an error.info
member… hopefully nothing too bad?