Add error handling for main loop
The reporting of errors at the end of the script is a crucial component,
and we must do our best to ensure that the script does not die before
reaching that point. Wrap the whole main loop in a big try-catch, and
also record the last item ID so that we know what we last worked on.