Add name in kana (P1814) to ja description for ja names
In Japanese, multiple different names can have the same transliteration
in kanji (“native label” property, P1705). To avoid errors from Wikibase
about multiple items having the same label and description in a
language, we need to add the unambiguous kana transliteration (“name in
kana” property, P1814) to the description as well.

If no such property is set, we show an error, with a new translatable
message; since no-P1705 and no-P282 already are identical save for the
property ID in all languages, I’ve taken the liberty to reuse those
translations for no-P1814 as well, without consulting with native
speakers, and hope that it’s correct.