Improve descriptions for kanji names
The previous code replaced the kanji with the kana for the description,
but the kana is not unique either, so we started getting errors for
ambiguous descriptions from Wikibase. Instead, put both in the
description, separated by a comma – except in Japanese, where the kanji
is already the label, so putting it in the description is redundant.

The comma is hard-coded for now, but we really should use MediaWiki’s
comma-separator message for it, as in some languages it should be a
full-width comma with no space after it.