Update dependencies
toml 3.0.0 is missing a changelog entry, but the only change appears to
be to some details regarding “constructor”-named properties, which
shouldn’t concern us. The only breaking change in mwbot 2.0.0 appears to
be dropping support for ancient Node.js and MediaWiki versions.

Also start tracking package-lock.json in Git so we know which package
versions worked last time whenever we come back to this years later. I’m
not sure why it was included in the .gitignore (9fb3370888 doesn’t say).