WIP: Better mul support
Trying to make namescript do something sensible in the face of mul. This
code is meant to ensure that, in all the language codes included in the
langlist (from namescript-data.js), the item will show its name as the
label, with as few actually saved labels as possible – ideally, it’ll
all just fall back to mul, but if someone has set a different label
somewhere in the fallback chain, we may have to set another label so it
won’t fall back to that other language. (Namescript doesn’t edit
existing labels, nor should it remove them if they’re different from the
name.)

namescript-lib.js just has some rudimentary code to fetch the fallback
languages (which we’ll need) and a pseudo-code sketch of what I think
the code should do. wip.js has an attempted implementation of that
pseudo-code, along with hard-coded copies of languageinfo and langlist,
so it can be tested more easily. Once I’m relatively confident it seems
to do the right thing for various inputs, I’ll start integrating it into
namescript-lib.js.