Integrate code into namescript-lib.js and add tests
Instead of manually testing this code, let’s start writing actual unit
tests for it, using the trusty Mocha+Chai combo. (Chai doesn’t like
being require()d these days, so the test file is .mjs and uses import
syntax. It’s fine.) Put prepareLabels() into namescript-lib.js where it
eventually belongs, and attach it to the namescript global so it’s
reachable for the tests (but don’t “advertise” it in the stub at the
top, unlike start() and translate()). Hard-code the languageinfo to make
the test resiliant against MediaWiki changes (and also faster), but get
the langlist from namescript-data.json, we don’t need to lug a copy of
that around forever. Write a few tests for the functionality I manually
tested so far, including one for the TODO mentioned in the previous
commit (test currently fails). Even more tests should be added later.