Add failing test
If lsb/hsb (Lower and Upper Sorbian) fall back to each other, but
subsequently fall back to de (German), and the de label differs from the
name, then it’s incorrect to remove the lsb/hsb labels. The code
currently does this; I need to fix this. (The test case just encodes the
last FIXME in the current code – turns out my suspicion was correct.)