Fix cycle with fallback to other label
To fix the dsb/hsb/de situation described in the previous commit, we
need to continue looking at the fallback chain, just making sure that
we’re not going to run the “will fall back all the way to mul” block at
the end. A boolean flag seems to work well enough for that.