Validate “cycle between fallback languages” handling
This code near the end of prepareLabels() is “dangerous”, in a way: its
existence can very easily mask bugs in the rest of the function, by
still removing any labels in the TODO list, even though the TODO list
should be almost empty at this point. This means that tests which check
that labels are removed might appear to work only due to this code,
turning it from a fix for a minor edge case into the “load bearing” part
of the function – this was previously seen in commit 14c0e8a692, and
now happened again (bug introduced in commit 040d63b2a3).

Avoid this trap by adding checks: if the TODO list contains anything
other than a cycle between fallback languages that are all set to the
name (i.e. if any fallback language is not set to be removed – either as
part of the codesToRemove or as part of the todoList – or has a label
different from the name), crash out of the whole function with an Error.

This revealed the bug introduced in commit 040d63b2a3: by including 'en'
in the list of fallback languages, we broke the main “todos” loop for
'en' as the languageCode itself, which now looked like it fell back to a
language ('en') where it wasn’t known if it would be removed or not (the
knownToFallBackToMul = false case); this in turn meant any language
falling back to en (that is: any language) couldn’t safely be removed in
the main “todos” loop either. Fix it by simply skipping the “fallbacks”
loop in this case (or, hypothetically, if any MediaWiki language code
included itself in its list of fallbacks, though I doubt that happens).