Make prepareLabels() return three sets of language codes
This way, prepareStuff() will be able to assemble a good edit summary,
counting the number of labels that will be set and removed. (I doubt
it’ll use codesToKeep, that’s mainly for the tests.) I don’t mind the
way the test expectations look now, either.

The added loop to delete codes to remove that don’t actually occur in
the labels is mainly important because without it the tests would have
to assert like two hundred expected codes to remove – otherwise we could
do that part in prepareStuff() when actually modifying the entity. And
yes, I did check if modifying a Set while iterating over it is safe, and
it does the right thing and is thoroughly specified in ECMAScript.