Add failing test for nb/nn/no
I noticed this when trying out the browser version of Namescript on
Q12325000, which has labels in Norwegian Bokmål (nb) and Norwegian
Nynorsk (nn). Both languages fall back to Norwegian (no), but that
language code is discouraged because it’s ambiguous between the other
two, so it’s not in the langlist. This means the analysis code gets
confused, and the test fails; I’ll have to figure out how to fix it.