Add langlistSet for faster membership check
The !langlistSet.has(fallbackCode) condition in the main loop is going
to be evaluated reasonably often, so let’s optimize it a bit.

There’s one .includes() call left, which is in the second (“cycle
cleanup”) loop, but at that point the workList should be almost empty,
so I don’t think it’s necessary to optimize that one. (Nor is it very
feasible, given that we’re checking membership in a list we’re also
modifying.)