Check that mul label doesn’t differ from P1705 value
If the mul label already exists and is different from the “native label”
statement, we shouldn’t change it, just like we don’t change other
existing labels. Add this as a separate check at the beginning with its
own error message; the error message phrasing is my attempt to *not*
imply that users should just copy either value to the other place
without thinking about it.